Add an explicit follow-through phase
This used to be done by `Land` angular velocity detection, but let's be
honest, that was pretty sloppy.

This fixes a bug where if you clip something during a dodge, you angular
velocity would momentarily dip, and `Land` would not be able to detect
you were dodging, so it would try to boost downwards or air-roll or do
all sorts of other crazy things.